feat(api): Make api directory contain all files (except .db) for api, move routes into new routes sub-directory
