+++ src/minitwit/api/helpers/auth.rb
