+++ src/minitwit/api/helpers/dbhelper.rb
