+++ src/minitwit/api/helpers/format.rb
