+++ src/minitwit/api/helpers/init.rb
