+++ src/minitwit/api/minitwit_api.rb
@@ -8,7 +8,7 @@ require 'securerandom'
 class MiniTwit < Sinatra::Base
   SECRET_KEY = SecureRandom.hex(32)
   PER_PAGE = 30
-  DATABASE = './db/minitwit.db'
+  DATABASE = '../db/minitwit.db'
   
   # Configure session management inside a configure block
   configure do
@@ -20,7 +20,9 @@ class MiniTwit < Sinatra::Base
   # Before request
   before do
     @db = connect_db(DATABASE)
-    @user = session[:user_id] ? query_db('SELECT * FROM user WHERE user_id = ?', [session[:user_id]], true) : nil
+    #ph = BCrypt::Password.create(params[:pwd]).to_s
+    #@user = query_db('SELECT * FROM user WHERE username = ? AND pw_hash = ?', [params[:username], ph], true)
+
     # Parse JSON request body
     begin
       content_type :json
@@ -39,7 +41,7 @@ class MiniTwit < Sinatra::Base
 
   # Note that 'helpers' is sinatra way of naming 'utils'
   require_relative 'helpers/init'
-  require_relative 'api/init'
+  require_relative 'routes/init'
 
   # Start the application
   run! if __FILE__ == $PROGRAM_NAME
