+++ src/minitwit/api/routes/fllws.rb
@@ -10,18 +10,19 @@ class MiniTwit < Sinatra::Base
 
     # halt 401, "Unauthorized" unless @user
 
-    whom_id = get_user_id(@data['follow']) || get_user_id(@data['unfollow'])
-    if whom_id.nil?
+    who_id = params[:username]
+    whom_id = @data['follow'] || @data['unfollow']
+    if whom_id.nil? || who_id.nil?
       status 404
-      body JSON({ 'message': "User #{@data['follow']} not found" })
+      body JSON({ 'message': "Either #{who_id} or #{@data} not found" })
     end
       
     if @data.key?('unfollow')
-      query_db('DELETE FROM follower WHERE who_id = ? AND whom_id = ?', [@data['username'], whom_id])  
+      query_db('DELETE FROM follower WHERE who_id = ? AND whom_id = ?', [who_id, whom_id])
       status 200
       body JSON({ 'message': 'Unfollowed #{whom_id}' })
     elsif @data.key?('follow')
-      query_db('INSERT INTO follower (who_id, whom_id) VALUES (?, ?)', [@data['username'], whom_id])
+      query_db('INSERT INTO follower (who_id, whom_id) VALUES (?, ?)', [who_id, whom_id])
       status 200
       body JSON({ 'message': 'Followed #{whom_id}',  })
     end
@@ -35,7 +36,7 @@ class MiniTwit < Sinatra::Base
 
     no_followers = request.env['no'] || 100
     followers = query_db('SELECT u.username FROM user u
-            INNER JOIN follower f on f.whom_id=u.user_id
+            INNER JOIN follower f on f.whom_id=u.username
             WHERE f.who_id = ?
             LIMIT ?', [params[:username], no_followers])
     follower_names = followers.map{|x| x['username']}
