+++ src/minitwit/api/routes/register.rb
